src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
aiohttp==3.8.4
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.8.3
    # via twilio
aiosignal==1.3.1
    # via aiohttp
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r src/requirements/requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   face-recognition
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cmake==3.26.3
    # via -r src/requirements/requirements.in
colorama==0.4.6
    # via click
dj-database-url==2.0.0
    # via -r src/requirements/requirements.in
django==4.1.9
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
dlib==19.24.1
    # via
    #   -r src/requirements/requirements.in
    #   face-recognition
face-recognition==1.3.0
    # via -r src/requirements/requirements.in
face-recognition-models==0.3.0
    # via face-recognition
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
kombu==5.2.4
    # via celery
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.24.3
    # via
    #   face-recognition
    #   opencv-contrib-python
    #   opencv-python
opencv-contrib-python==4.7.0.72
    # via -r src/requirements/requirements.in
opencv-python==4.7.0.72
    # via -r src/requirements/requirements.in
pillow==9.5.0
    # via
    #   -r src/requirements/requirements.in
    #   face-recognition
prompt-toolkit==3.0.38
    # via click-repl
pyjwt==2.7.0
    # via twilio
python-decouple==3.8
    # via -r src/requirements/requirements.in
pytz==2023.3
    # via
    #   celery
    #   twilio
redis==4.5.5
    # via -r src/requirements/requirements.in
requests==2.31.0
    # via twilio
six==1.16.0
    # via click-repl
sqlparse==0.4.4
    # via django
twilio==8.2.1
    # via -r src/requirements/requirements.in
typing-extensions==4.6.2
    # via dj-database-url
tzdata==2023.3
    # via django
urllib3==2.0.2
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
yarl==1.9.2
    # via aiohttp
